reworked shopping list component
